Refactorizacion de Pelicula
Se modificaron el repositorio, servicio y controlador de pelicula
